Add log arithmetic
Turn the original way of calculating log into self-defined way of
calculating log with fixed-point arithmetic.

The precision of UQ23.8 is insufficient, so I reset the type of the parameter score to unsigned long (64 bits) and adjusted frac_bits to 16 bits.

+++ game.h
@@ -21,5 +21,5 @@ extern const line_t lines[4];
 
 int *available_moves(const char *table);
 char check_win(char *t);
-unsigned calculate_win_value(char win, char player);
+unsigned long calculate_win_value(char win, char player);
 void draw_board(const char *t);
